+++ src/environments/constants.ts
@@ -7,9 +7,9 @@ const codeValue = {
   queryTaskScope: {
     sqlQuery: 'sql-query',
     webApiQuery: 'web-api-query',
+    webApiQueryNoProxy: 'web-api-query-no-proxy',
     cartographyQuery: 'cartography-query',
-    urlQuery: 'URL',
-    resource: 'resource'
+    urlQuery: 'external-link'
   },
   editionTaskScope: {
     dbEdition: 'db-edit',